add jwt auth

+++ src/chat/chat.controller.ts
@@ -1,5 +1,4 @@
 import {
-  BadRequestException,
   Body,
   Controller,
   Delete,
@@ -9,6 +8,9 @@ import {
   Param,
   Patch,
   Post,
+  Request,
+  UnauthorizedException,
+  UseGuards,
 } from '@nestjs/common';
 import { ChatService } from './chat.service';
 import { ApiBody, ApiResponse } from '@nestjs/swagger';
@@ -16,19 +18,34 @@ import CreateChatDto from 'src/dto/create-chat.dto';
 import { CreateMessageDto } from 'src/dto/create-message.dto';
 import CreateMemberDto from 'src/dto/create-member.dto';
 import CreateRemoveMemberDto from 'src/dto/create-remove-member.dto';
+import JwtAuthGuard from 'src/auth/jwt-auth.guard';
+import { MessageService } from './message.service';
 
 @Controller('chat')
 export class ChatController {
-  constructor(private readonly chatService: ChatService) {}
+  constructor(
+    private readonly chatService: ChatService,
+    private readonly messageService: MessageService,
+  ) {}
 
   @Get(':chatId')
+  @UseGuards(JwtAuthGuard)
   @ApiResponse({ status: 200, description: 'OK' })
   @ApiResponse({ status: 400, description: 'Chat with this id is not exist' })
-  async getChat(@Param('chatId') chatId: number) {
+  async getChat(@Request() req, @Param('chatId') chatId: number) {
+    if (
+      !(await this.chatService.isUserInMembers({
+        chatId: chatId,
+        userId: req.user.id,
+      }))
+    )
+      throw new UnauthorizedException();
+
     return await this.chatService.getChat(chatId);
   }
 
   @Post()
+  @UseGuards(JwtAuthGuard)
   @HttpCode(HttpStatus.CREATED)
   @ApiResponse({ status: 201, description: 'Created' })
   @ApiResponse({
@@ -39,11 +56,15 @@ export class ChatController {
     status: 400,
     description: `User with this id already in that chat`,
   })
-  async createChat(@Body() chat: CreateChatDto) {
+  async createChat(@Request() req, @Body() chat: CreateChatDto) {
+    if ((await chat.members.findIndex((elem) => elem === req.user.id)) === -1)
+      await chat.members.push(req.user.id);
+
     return await this.chatService.createChat(chat);
   }
 
   @Patch('members')
+  @UseGuards(JwtAuthGuard)
   @ApiBody({ type: CreateMemberDto })
   @HttpCode(HttpStatus.CREATED)
   @ApiResponse({ status: 201, description: 'Created' })
@@ -55,7 +76,18 @@ export class ChatController {
     status: 400,
     description: `User with this id already in that chat`,
   })
-  async addMemberToChat(@Body() { users, chatId }: CreateMemberDto) {
+  async addMembersToChat(
+    @Request() req,
+    @Body() { users, chatId }: CreateMemberDto,
+  ) {
+    if (
+      !(await this.chatService.isUserInMembers({
+        chatId: chatId,
+        userId: req.user.id,
+      }))
+    )
+      throw new UnauthorizedException();
+
     await users.forEach(async (elem) => {
       await this.chatService.addMember(elem, chatId);
     });
@@ -63,6 +95,7 @@ export class ChatController {
   }
 
   @Delete('members')
+  @UseGuards(JwtAuthGuard)
   @HttpCode(HttpStatus.NO_CONTENT)
   @ApiBody({ type: CreateRemoveMemberDto })
   @ApiResponse({ status: 204 })
@@ -75,41 +108,57 @@ export class ChatController {
     description: 'Chat with this id is not exist',
   })
   async removeMemberFromChat(
+    @Request() req,
     @Body() { userId, chatId }: CreateRemoveMemberDto,
   ) {
-    return await this.chatService.removeMember(userId, chatId);
+    return await this.chatService.removeMember(
+      userId ? userId : req.user.id,
+      chatId,
+    );
   }
 
   @Delete(':chatId')
+  @UseGuards(JwtAuthGuard)
   @HttpCode(HttpStatus.NO_CONTENT)
   @ApiResponse({ status: 204 })
   @ApiResponse({
     status: 400,
     description: `Chat with this id in not exitst`,
   })
-  async deleteChat(@Param('chatId') chatId: string) {
+  async deleteChat(@Request() req, @Param('chatId') chatId: string) {
+    if (
+      !(await this.chatService.isUserInMembers({
+        chatId: +chatId,
+        userId: req.user.id,
+      }))
+    )
+      throw new UnauthorizedException();
     await this.chatService.removeChat(+chatId);
   }
 
   @Post('message/')
+  @UseGuards(JwtAuthGuard)
   @HttpCode(HttpStatus.CREATED)
   @ApiBody({ type: CreateMessageDto })
   @ApiResponse({ status: 201, description: 'Message created' })
-  async createChatMessage(@Body() message: CreateMessageDto) {
+  async createChatMessage(@Request() req, @Body() message: CreateMessageDto) {
     if (
       !(await this.chatService.isUserInMembers({
-        userId: +message.user,
         chatId: +message.chat,
+        userId: req.user.id,
       }))
     )
-      throw new BadRequestException('This user is not in members of this chat');
+      throw new UnauthorizedException();
 
-    const dbMessage = await this.chatService.createMessage(message);
+    message.user = req.user.id;
 
-    return await this.chatService.addMessageToChat(dbMessage);
+    const dbMessage = await this.messageService.createMessage(message);
+
+    return await this.messageService.addMessageToChat(dbMessage);
   }
 
   @Patch('message/:messageId')
+  @UseGuards(JwtAuthGuard)
   @ApiBody({ type: CreateMessageDto })
   @ApiResponse({
     status: 200,
@@ -119,20 +168,26 @@ export class ChatController {
     description: `Message with this id is not exist`,
   })
   async editMessage(
+    @Request() req,
     @Param('messageId') messageId: string,
     @Body() message: Partial<CreateMessageDto>,
   ) {
-    return await this.chatService.updateMessage(+messageId, message);
+    if (!(await this.messageService.isMessageInUser(req.user.id, +messageId)))
+      throw new UnauthorizedException();
+    return await this.messageService.updateMessage(+messageId, message);
   }
 
   @Delete('message/:messageId')
+  @UseGuards(JwtAuthGuard)
   @HttpCode(HttpStatus.NO_CONTENT)
   @ApiResponse({ status: 204, description: 'Ok' })
   @ApiResponse({
     status: 400,
     description: 'Message with this ida is not exist',
   })
-  async removeMessage(@Param('messageId') messageId: string) {
-    return await this.chatService.deleteMessage(+messageId);
+  async removeMessage(@Request() req, @Param('messageId') messageId: string) {
+    if (!(await this.messageService.isMessageInUser(req.user.id, +messageId)))
+      throw new UnauthorizedException();
+    return await this.messageService.deleteMessage(+messageId);
   }
 }
